@@ -7,62 +7,62 @@
     "build": "nx build web-yardbird",
     "test": "nx test",
     "clean": "npx rimraf hooks node_modules package-lock.json yarn.lock && yarn",
-    "postinstall": "ngcc --properties es2015 browser module main"
+    "postinstall": "ngcc --properties es2020 browser module main"
   },
   "private": true,
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.0.0",
-    "@angular-eslint/eslint-plugin": "~13.0.0",
-    "@angular-eslint/eslint-plugin-template": "~13.0.0",
-    "@angular-eslint/template-parser": "~13.0.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "^13.0.0",
-    "@angular/language-service": "^13.0.0",
+    "@angular-devkit/build-angular": "14.0.1",
+    "@angular-eslint/eslint-plugin": "13.2.1",
+    "@angular-eslint/eslint-plugin-template": "13.2.1",
+    "@angular-eslint/template-parser": "13.2.1",
+    "@angular/cli": "~14.0.0",
+    "@angular/compiler-cli": "14.0.1",
+    "@angular/language-service": "14.0.1",
     "@nativescript/eslint-plugin": "~0.0.4",
     "@nativescript/types": "~8.2.0",
     "@nativescript/webpack": "~5.0.0",
-    "@ngtools/webpack": "^13.0.0",
-    "@nrwl/angular": "^13.0.0",
+    "@ngtools/webpack": "^14.0.0",
+    "@nrwl/angular": "14.3.1",
     "@nrwl/cli": "13.9.5",
-    "@nrwl/cypress": "13.10.4",
-    "@nrwl/eslint-plugin-nx": "13.9.5",
-    "@nrwl/jest": "13.9.5",
-    "@nrwl/linter": "13.9.5",
-    "@nrwl/workspace": "13.9.5",
-    "@nstudio/angular": "13.4.3",
-    "@nstudio/nativescript": "13.4.3",
-    "@nstudio/nativescript-angular": "13.4.3",
-    "@nstudio/web": "13.4.3",
-    "@nstudio/web-angular": "13.4.3",
-    "@nstudio/xplat": "^13.4.3",
-    "@types/jest": "27.0.2",
+    "@nrwl/cypress": "14.3.1",
+    "@nrwl/eslint-plugin-nx": "14.3.1",
+    "@nrwl/jest": "14.3.1",
+    "@nrwl/linter": "14.3.1",
+    "@nrwl/workspace": "14.3.1",
+    "@nstudio/angular": "14.0.1",
+    "@nstudio/nativescript": "14.0.1",
+    "@nstudio/nativescript-angular": "14.0.1",
+    "@nstudio/web": "14.0.1",
+    "@nstudio/web-angular": "14.0.1",
+    "@nstudio/xplat": "^14.0.1",
+    "@types/jest": "27.4.1",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "~5.10.0",
-    "@typescript-eslint/parser": "~4.33.0",
+    "@typescript-eslint/eslint-plugin": "5.24.0",
+    "@typescript-eslint/parser": "5.24.0",
     "cypress": "^9.1.0",
-    "eslint": "8.2.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.1.1",
-    "nx": "13.9.5",
-    "prettier": "^2.5.1",
+    "jest": "27.5.1",
+    "jest-preset-angular": "11.1.2",
+    "nx": "14.3.1",
+    "prettier": "2.6.2",
     "sass": "^1.30.0",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.5.2"
+    "ts-jest": "27.1.4",
+    "typescript": "4.7.3"
   },
   "dependencies": {
-    "@angular/animations": "^13.0.0",
-    "@angular/cdk": "^13.3.7",
-    "@angular/common": "^13.0.0",
-    "@angular/compiler": "^13.0.0",
-    "@angular/core": "^13.0.0",
+    "@angular/animations": "14.0.1",
+    "@angular/cdk": "14.0.1",
+    "@angular/common": "14.0.1",
+    "@angular/compiler": "14.0.1",
+    "@angular/core": "14.0.1",
     "@angular/fire": "^7.3.0",
-    "@angular/forms": "^13.0.0",
-    "@angular/platform-browser": "^13.0.0",
-    "@angular/platform-browser-dynamic": "^13.0.0",
-    "@angular/router": "^13.0.0",
-    "@nativescript/angular": "^13.0.0",
+    "@angular/forms": "14.0.1",
+    "@angular/platform-browser": "14.0.1",
+    "@angular/platform-browser-dynamic": "14.0.1",
+    "@angular/router": "14.0.1",
+    "@nativescript/angular": "^14.0.0-rc.0",
     "@nativescript/core": "^8.2.0",
     "@nativescript/firebase-auth": "^2.0.1",
     "@nativescript/firebase-core": "^2.0.1",
