@@ -1,14 +1,16 @@
-module.exports = {
-  displayName: 'xplat-web-features',
+/* eslint-disable */
+export default {
+  displayName: 'xplat-nativescript-core',
   preset: '../../../../jest.preset.js',
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/tsconfig.spec.json',
     },
   },
+  testEnvironment: 'node',
   transform: {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../../coverage/libs/xplat/web/features',
+  coverageDirectory: '../../../../coverage/libs/xplat/nativescript/core',
 };
