@@ -1,6 +1,7 @@
-module.exports = {
-  displayName: 'xplat-nativescript-utils',
-  preset: '../../../../jest.preset.js',
+/* eslint-disable */
+export default {
+  displayName: 'xplat-utils',
+  preset: '../../../jest.preset.js',
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/tsconfig.spec.json',
@@ -11,5 +12,5 @@ module.exports = {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../../coverage/libs/xplat/nativescript/utils',
+  coverageDirectory: '../../../coverage/libs/xplat/utils',
 };
