@@ -1,6 +1,7 @@
-module.exports = {
-  displayName: 'xplat-scss',
-  preset: '../../../jest.preset.js',
+/* eslint-disable */
+export default {
+  displayName: 'xplat-web-features',
+  preset: '../../../../jest.preset.js',
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/tsconfig.spec.json',
@@ -10,5 +11,5 @@ module.exports = {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../coverage/libs/xplat/scss',
+  coverageDirectory: '../../../../coverage/libs/xplat/web/features',
 };
