@@ -1,15 +1,15 @@
-module.exports = {
-  displayName: 'xplat-features',
-  preset: '../../../jest.preset.js',
+/* eslint-disable */
+export default {
+  displayName: 'xplat-web-scss',
+  preset: '../../../../jest.preset.js',
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/tsconfig.spec.json',
     },
   },
-  testEnvironment: 'node',
   transform: {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../coverage/libs/xplat/features',
+  coverageDirectory: '../../../../coverage/libs/xplat/web/scss',
 };
