@@ -1,3 +1,4 @@
-const nxPreset = require('@nrwl/jest/preset');
+const nxPreset = require('@nrwl/jest/preset').default;
+
 
 module.exports = { ...nxPreset };
