@@ -1,9 +1,11 @@
+import { inject } from "@angular/core";
 import { ItemService } from "../services/item.service";
 
 export abstract class DetailsBaseComponent {
   item$;
+  #itemService = inject(ItemService);
 
-  constructor(private itemService: ItemService) {
-    this.item$ = this.itemService.item$;
+  constructor() {
+    this.item$ = this.#itemService.item$;
   }
 }
